chore: divide memoRoom controller
